requirements.txt
asgiref==3.6.0
click==8.1.3
fastapi==0.68.1
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.7
python-dotenv==1.0.0
starlette==0.14.2
typing_extensions==4.5.0
uvicorn==0.15.0
jwt
jose
mysql-connector-python
pyodbc
pandas
openpyxl
matplotlib
httpx
requests
paramiko
pycryptodome